// src/lib/api/apiSchema.generated.ts
export const API_ENUMS = {
// ...
  "BookingFormPartyDto": {
    "party_kind": [
      "SHIPPER",
      "CONSIGNEE",
      "NOTIFY",
      "BILLING",
      "AGENT"
    ],
    "entity_kind": [
      "COMPANY",
      "INDIVIDUAL"
    ]
  },
// ...
  "UpsertNvoccBookingFormDto": {
    "service_scope": [
      "DOOR_TO_DOOR",
      "DOOR_TO_PORT",
      "PORT_TO_DOOR",
      "PORT_TO_PORT"
    ],
    "activity_sector": [
      "CIVILIAN",
      "MILITARY",
      "NUCLEAR"
    ]
  },
// ...
} as const;